Add support for cfgfs session interface
This is an optional interface that displays info about all sessions and
not just the lead session on an ACL or the initiator name for dynamic
sessions.

To tell the kernel that userspace supports the new interface, we have to
make the root "sessions" dir in the ACL dir or TPG dir. Those dirs will
then have a dir for each session that is accessed through that object
similar to how the kernel reports ACL based sessions under the ACL node
and dynamic sessions in the TPG dir. The name of the dir is
"session-$SID".

During target removal we then must delete the sessions under the parent
object like is done for other children, but because the kernel made the
session there is no mkdir/rmdir support so we write the $SID to the
tpgt/sessions/remove file.